latest working config

+++ spring/eshop/src/main/java/it/akademija/repositories/ProductRepository.java
@@ -3,7 +3,6 @@ package it.akademija.repositories;
 
 import it.akademija.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
